Make nodes handle their respective navigation source geometry
Makes nodes handle their respective navigation source geometry.
