+++ scene/2d/multimesh_instance_2d.h
@@ -34,6 +34,9 @@
 #include "scene/2d/node_2d.h"
 #include "scene/resources/multimesh.h"
 
+class NavigationPolygon;
+class NavigationMeshSourceGeometryData2D;
+
 class MultiMeshInstance2D : public Node2D {
 	GDCLASS(MultiMeshInstance2D, Node2D);
 
@@ -55,6 +58,14 @@ public:
 
 	void set_texture(const Ref<Texture2D> &p_texture);
 	Ref<Texture2D> get_texture() const;
+
+private:
+	static Callable _navmesh_source_geometry_parsing_callback;
+	static RID _navmesh_source_geometry_parser;
+
+public:
+	static void navmesh_parse_init();
+	static void navmesh_parse_source_geometry(const Ref<NavigationPolygon> &p_navigation_mesh, Ref<NavigationMeshSourceGeometryData2D> p_source_geometry_data, Node *p_node);
 
 	MultiMeshInstance2D();
 	~MultiMeshInstance2D();
