+++ scene/2d/polygon_2d.h
@@ -33,6 +33,9 @@
 
 #include "scene/2d/node_2d.h"
 
+class NavigationPolygon;
+class NavigationMeshSourceGeometryData2D;
+
 class Polygon2D : public Node2D {
 	GDCLASS(Polygon2D, Node2D);
 
@@ -149,6 +152,14 @@ public:
 
 	void set_skeleton(const NodePath &p_skeleton);
 	NodePath get_skeleton() const;
+
+private:
+	static Callable _navmesh_source_geometry_parsing_callback;
+	static RID _navmesh_source_geometry_parser;
+
+public:
+	static void navmesh_parse_init();
+	static void navmesh_parse_source_geometry(const Ref<NavigationPolygon> &p_navigation_mesh, Ref<NavigationMeshSourceGeometryData2D> p_source_geometry_data, Node *p_node);
 
 	Polygon2D();
 	~Polygon2D();
