+++ scene/3d/multimesh_instance_3d.h
@@ -34,6 +34,9 @@
 #include "scene/3d/visual_instance_3d.h"
 #include "scene/resources/multimesh.h"
 
+class NavigationMesh;
+class NavigationMeshSourceGeometryData3D;
+
 class MultiMeshInstance3D : public GeometryInstance3D {
 	GDCLASS(MultiMeshInstance3D, GeometryInstance3D);
 
@@ -53,6 +56,14 @@ public:
 	Array get_meshes() const;
 
 	virtual AABB get_aabb() const override;
+
+private:
+	static Callable _navmesh_source_geometry_parsing_callback;
+	static RID _navmesh_source_geometry_parser;
+
+public:
+	static void navmesh_parse_init();
+	static void navmesh_parse_source_geometry(const Ref<NavigationMesh> &p_navigation_mesh, Ref<NavigationMeshSourceGeometryData3D> p_source_geometry_data, Node *p_node);
 
 	MultiMeshInstance3D();
 	~MultiMeshInstance3D();
