+++ scene/3d/physics/static_body_3d.h
@@ -33,6 +33,9 @@
 
 #include "scene/3d/physics/physics_body_3d.h"
 
+class NavigationMesh;
+class NavigationMeshSourceGeometryData3D;
+
 class StaticBody3D : public PhysicsBody3D {
 	GDCLASS(StaticBody3D, PhysicsBody3D);
 
@@ -59,6 +62,13 @@ public:
 
 private:
 	void _reload_physics_characteristics();
+
+	static Callable _navmesh_source_geometry_parsing_callback;
+	static RID _navmesh_source_geometry_parser;
+
+public:
+	static void navmesh_parse_init();
+	static void navmesh_parse_source_geometry(const Ref<NavigationMesh> &p_navigation_mesh, Ref<NavigationMeshSourceGeometryData3D> p_source_geometry_data, Node *p_node);
 };
 
 #endif // STATIC_BODY_3D_H
